Fix panel admin ajout

+++ res/php/admin/supprimer.php
@@ -7,10 +7,10 @@ $co = mysqli_connect(DB_HOST, DB_USERNAME, DB_PASSWORD, DB_NAME);
 $idT=mysqli_escape_string($co, htmlspecialchars($_GET['idT']));
 $table= mysqli_escape_string($co, htmlspecialchars($_GET['table']));
 
-// if (!isset($_SESSION['utilisateur']) || !($_SESSION['utilisateur'] instanceof Utilisateur)) {
-//     header("Location: ./index.php");
-//     exit();
-// } 
+if (!isset($_SESSION['utilisateur']) || !($_SESSION['utilisateur'] instanceof Utilisateur) || !($_SESSION['utilisateur']->getRole())) {
+    header("Location: ../../../index.php");
+    exit();
+} 
 
 if ($table=='tipsEco'){
     $req = $co->query("DELETE FROM $table WHERE idTips=$idT");
